// frontend/src/types/patient-registration.ts
export type IPatientRegistrationItem = {
  id: number;
  recordId: string;
  uhid: string;
  registrationDate: string;
  caseDate: string;
  firstName: string;
  middleName: string | null;
  lastName: string;
  category: string;
  birthDate: string | null;
  ageUnit: string | null;
  age: number | null;
  sex: string;
  weddingDate?: string | null;
  hospitalDoctor: string;
  referenceDoctor: string;
  language: string;
  religion: string;
  height: number | null;
  weight: number | null;
  bloodGroup: string | null;
  maritalStatus: string | null;
  MaritalStatus: string | null;
  mobile: string;
  mobile2: string | null;
  office: string | null;
  email: string | null;
  address: string | null;
  area: string | null;
  city: string | null;
  state: string | null;
  pin: string | null;
  nationality: string;
  typeOfVisa: string | null;
  estimate: string | null;
  adharCard: string | null;
  profileImage: string | undefined;
  panCard: string | null;
  membershipId: string | null;
  employeesId: string | null;
  occupation: string | null;
  spouseOccupation: string | null;
  companyName: string | null;
  education: string | null;
  yojanaCardNo: string | null;
  mediclaim: boolean;
  remark: string | null;
  referralTo: string | null;
  transfer: string | null;
  createdByAdminUser?: IUserItem;
  createdByUser?: IUserItem;
  updatedByAdminUser?: IUserItem;
  updatedByUser?: IUserItem;
  createdAt: string;
  updatedAt: string;
};